# src/routes/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Viaro's Nodejs",
        "description": "Viaro's Nodejs training project"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local Development environment"
        }
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/donation/saveDonation": {
            "post": {
                "tags": [
                    "Donation"
                ],
                "summary": "Create a donation",
                "requestBody": {
                    "description": "Donation model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/donation"
                            }
                        }
                    }
                },
                "produces": [
                    "text/html; charset=utf-8"
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/donation/getDonations/{userid}": {
            "get": {
                "tags": [
                    "Donation"
                ],
                "summary": "Create a donation",
                "parameters": [
                    {
                        "in": "path",
                        "name": "userid",
                        "required": true,
                        "description": "User's Id",
                        "schema": {
                            "$ref": "#/definitions/userid"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/donationResponse"
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
        "/donation/getDonations/worksheet/{userid}": {
            "get": {
                "tags": [
                    "Donation"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userid",
                        "required": true,
                        "description": "User's Id",
                        "schema": {
                            "$ref": "#/definitions/userid"
                        }
                    }
                ],
                "summary": "Get excel file with user donations",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "definitions": {
        "userid": {
            "type": "string"
        },
...
        "donationResponse": {
            "type": "object",
            "properties": {
                "Date": {
                    "type": "string"
                },
                "Amount": {
                    "type": "number"
                },
                "AgencyName": {
                    "type": "string"
                },
                "Frequency": {
                    "type": "number"
                },
                "Country": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                }
            }
        },
...
        "donation": {
            "type": "object",
            "properties": {
                "Frequency": {
                    "type": "number"
                },
                "Amount": {
                    "type": "number"
                },
                "CardType": {
                    "type": "string"
                },
                "CardNumber": {
                    "type": "string"
                },
                "CVV": {
                    "type": "string"
                },
                "NameCard": {
                    "type": "string"
                },
                "ExpirationDate": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "Country": {
                    "type": "string"
                },
                "Address1": {
                    "type": "string"
                },
                "Address2": {
                    "type": "string"
                },
                "City": {
                    "type": "string"
                },
                "State": {
                    "type": "string"
                },
                "ZipCode": {
                    "type": "string"
                },
                "AgencyId": {
                    "type": "string"
                },
                "Date": {
                    "type": "string"
                },
                "UserId": {
                    "type": "string"
                }
            }
        }
    }
}